+++ CHANGELOG.md
@@ -1,5 +1,30 @@
 ## Next
 
+## 3.17.0 - 2025-06-05
+
+- chore: do not capture screenshot during screen changes ([#254](https://github.com/PostHog/posthog-android/pull/254))
+
+## 3.16.0 - 2025-05-27
+
+- chore: manual start and stop session replay ([#248](https://github.com/PostHog/posthog-android/pull/248))
+
+## 3.15.1 - 2025-05-27
+
+- fix: clear feature flags cache when flags were cleared up server side ([#246](https://github.com/PostHog/posthog-android/pull/246)) 
+
+## 3.15.0 - 2025-05-14
+
+- feat: log request and payload requests if debug is enabled ([#249](https://github.com/PostHog/posthog-android/pull/249))
+
+## 3.14.1 - 2025-04-23
+
+- fix: Send correct `$feature_flag_response` for the `$feature_flag_called` event when calling `isFeatureEnabled` ([#244](https://github.com/PostHog/posthog-android/pull/244))
+- feat: use network time instead of device time ([#235](https://github.com/PostHog/posthog-android/pull/235))
+
+## 3.14.0 - 2025-04-21
+
+- feat: add `$feature_flag_id`, `$feature_flag_version`, `$feature_flag_reason` and `$feature_flag_request_id` properties to `$feature_flag_called` event ([#243](https://github.com/PostHog/posthog-android/pull/243))
+
 ## 3.13.1 - 2025-04-01
 
 - fix: add remote config response to proguard rules ([#238](https://github.com/PostHog/posthog-android/pull/238))
@@ -305,7 +330,7 @@ Check out the [USAGE](https://github.com/PostHog/posthog-android/blob/main/USAGE
 ## 3.0.0-alpha.7 - 2023-10-10
 
 - isFeatureEnabled now returns true if multivariant flag ([#42](https://github.com/PostHog/posthog-android/pull/42))
-- getFeatureFlagPayload returns non strigified JSON ([#44](https://github.com/PostHog/posthog-android/pull/44))
+- getFeatureFlagPayload returns non stringified JSON ([#44](https://github.com/PostHog/posthog-android/pull/44))
 
 ## 3.0.0-alpha.6 - 2023-10-06